# swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "DEX aggregation and swap-quote API for EVM chains.\nMaintains a catalogue of blockchains, tokens, DEXes and liquidity\npools, and serves executable swap quotes built from on-chain AMM data.",
        "title": "uatu API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/quotes": {
            "post": {
                "description": "Resolves the pools for a token pair on the given chain, prices the swap against\neach DEX's on-chain contracts concurrently, and returns the best output together\nwith the encoded Permit2 approval and swap calldata needed to execute it.\nThe swap calldata is guarded by amountOutMinimum, the quoted output reduced by\nrequest's slippageBps (the server default when omitted), so a fill below it reverts.\nQuotes carry a 5-minute deadline and are persisted with a pending status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Create a Swap Quote",
                "parameters": [
                    {
                        "description": "request body to create a swap quote",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/uatu.QuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/server.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/uatu.Quote"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.APIResponse"
                        }
                    }
                }
            }
        },
        "/quotes/routes": {
            "post": {
                "description": "Prices a swap against every supported DEX and returns the valid routes ordered by output amount.\nEach route reports amountOutMinimum, the quoted output reduced by the request's\n(the server default when omitted), which is the floor the swap would be executed against.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Compare DEX Routes",
                "parameters": [
                    {
                        "description": "request body to compare DEX quotes",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/uatu.QuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/server.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/uatu.RouteQuote"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.APIResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "uatu.Actions": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "chainId": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "spender": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "uatu.Quote": {
            "type": "object",
            "properties": {
                "amountIn": {
                    "type": "string"
                },
                "amountInFloat": {
                    "type": "string"
                },
                "amountOut": {
                    "type": "string"
                },
                "amountOutFloat": {
                    "type": "string"
                },
                "amountOutMinimum": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deadline": {
                    "type": "integer"
                },
                "destinationChain": {
                    "type": "string"
                },
                "destinationChainId": {
                    "type": "integer"
                },
                "explorerUrl": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "originChain": {
                    "type": "string"
                },
                "originChainId": {
                    "type": "integer"
                },
                "pairAddress": {
                    "type": "string"
                },
                "quoteId": {
                    "type": "string"
                },
                "recipientAddress": {
                    "type": "string"
                },
                "route": {
                    "$ref": "#/definitions/uatu.Route"
                },
                "slippageBps": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "completed",
                        "failed"
                    ]
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/uatu.Actions"
                    }
                },
                "tokenIn": {
                    "$ref": "#/definitions/uatu.Token"
                },
                "tokenOut": {
                    "$ref": "#/definitions/uatu.Token"
                },
                "updatedAt": {
                    "type": "string"
                },
                "walletAddress": {
                    "type": "string"
                }
            }
        },
        "uatu.QuoteRequest": {
            "type": "object",
            "required": [
                "amount",
                "chain",
                "chainId",
                "recipientAddress",
                "tokenIn",
                "tokenOut"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "chain": {
                    "type": "string"
                },
                "chainId": {
                    "type": "integer"
                },
                "recipientAddress": {
                    "type": "string"
                },
                "slippageBps": {
                    "type": "integer"
                },
                "tokenIn": {
                    "type": "string"
                },
                "tokenOut": {
                    "type": "string"
                }
            }
        },
        "uatu.Route": {
            "type": "object",
            "properties": {
                "dex": {
                    "type": "string"
                },
                "fees": {
                    "description": "Fees is the route fee in the input token's smallest unit.",
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "uatu.RouteQuote": {
            "type": "object",
            "properties": {
                "amountIn": {
                    "type": "string"
                },
                "amountOut": {
                    "type": "string"
                },
                "amountOutMinimum": {
                    "type": "string"
                },
                "deadline": {
                    "type": "integer"
                },
                "route": {
                    "$ref": "#/definitions/uatu.Route"
                },
                "slippageBps": {
                    "type": "integer"
                },
                "tokenIn": {
                    "$ref": "#/definitions/uatu.Token"
                },
                "tokenOut": {
                    "$ref": "#/definitions/uatu.Token"
                }
            }
        },
        "uatu.Token": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "blockchainId": {
                    "type": "integer"
                },
                "decimals": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}